refactor: convert the lib with a CLI structure to a CLI app structure
