--- .ai/cli_interface.md
@@ -1,18 +1,0 @@
-# CLI Interface
-
-## Current (Phase 1 ✅)
-
-```bash
-cd packages/lib && pnpm cli
-```
-
-**Interactive Flow:**
-
-1. Select source from YAML config
-2. Configure options (maxPages only)
-3. Start crawling with real-time progress
-4. View results with pagination
-
-## Future (Phase 3)
-
-Command-based interface: `ethos crawl <source> --max-pages 5`
