--- .ai/scraping_details.md
@@ -1,22 +1,0 @@
-# Crawling Implementation (Phase 1 ✅)
-
-## ArticleListingCrawler Features
-
-- **Pagination**: Auto next-page navigation
-- **Deduplication**: Skip duplicate URLs
-- **Mandatory Content Extraction**: Always extracts content pages (no skip option)
-- **Error Recovery**: Continue on individual failures
-- **Field Stats**: Track extraction success rates
-
-## Error Handling
-
-- **Required fields**: Skip item, continue crawl
-- **Optional fields**: Set undefined, continue
-- **Page failures**: Log error, continue pagination
-- **Content failures**: Fall back to listing data (with warnings)
-
-## Future Crawlers
-
-- **RSS**: XML feed parsing
-- **API**: JSON endpoint integration
-- **Social**: Platform-specific extractors
