--- .ai/sources.md
@@ -1,20 +1,0 @@
-# Data Sources
-
-## Implemented ✅
-
-- **EFF Updates**: https://www.eff.org/updates (full config in `sources.yaml`)
-
-## Phase 2 Targets
-
-1. AccessNow - https://www.accessnow.org/news-updates/
-2. TorrentFreak - https://torrentfreak.com/
-3. Big Brother Watch - https://bigbrotherwatch.org.uk/media/
-4. Citizenlab - https://citizenlab.ca/2025/
-5. Declassified UK - https://www.declassifieduk.org/category/archive/
-6. Freedom Press - https://freedom.press/issues
-
-## Future Types
-
-- **RSS**: OONI Blog feed
-- **API**: GDELT Project
-- **Social**: Reddit, Mastodon, etc.
