+++ .gitignore
@@ -1,5 +1,5 @@
-# Lib
 storage/
+.ai
 
 # Dependencies
 node_modules/
