+++ package.json
@@ -2,20 +2,21 @@
 	"name": "ethos",
 	"version": "1.0.0",
 	"description": "Ethos crawls then serves digital rights organizations' publications through its API.",
-	"main": "dist/index.js",
-	"types": "dist/index.d.ts",
+	"bin": {
+		"ethos": "dist/index.js"
+	},
 	"scripts": {
 		"lint": "biome check .",
 		"format": "biome format --write .",
 		"commit": "git-cz",
 		"prepare": "lefthook install",
 		"watch": "tsc --watch",
-		"build": "tsc && tsc-alias",
+		"build": "vite build",
+		"start": "node dist/index.js",
 		"test": "vitest run",
 		"test:coverage": "vitest run --coverage",
 		"test:watch": "vitest --watch",
-		"clean": "rm -rf dist",
-		"cli": "node dist/cli/index.js"
+		"clean": "rm -rf dist"
 	},
 	"repository": {
 		"type": "git",
