--- src/cli/index.ts
@@ -1,29 +1,0 @@
-import { join } from "node:path";
-import { Command } from "commander";
-import {
-	ArticleListingCrawler,
-	CrawlerRegistry,
-	ProcessingPipeline,
-	SourceRegistry,
-} from "@/index.js";
-import { showMainMenu } from "./menu.js";
-
-const program = new Command();
-
-const sourceRegistry = new SourceRegistry(
-	join(process.cwd(), "src", "config", "sources.yaml"),
-);
-const crawlerRegistry = new CrawlerRegistry();
-const pipeline = new ProcessingPipeline(crawlerRegistry, "./storage");
-
-crawlerRegistry.register(new ArticleListingCrawler());
-
-program
-	.name("ethos")
-	.description("Ethos web crawling command line interface")
-	.version("1.0.0")
-	.action(async () => {
-		await showMainMenu(sourceRegistry, pipeline);
-	});
-
-program.parseAsync(process.argv).catch(console.error);
