+++ src/ui/display.ts
@@ -1,4 +1,7 @@
-import type { ProcessingPipeline, ProcessingSummaryResult } from "@/index.js";
+import type {
+	ProcessingPipeline,
+	ProcessingSummaryResult,
+} from "../core/ProcessingPipeline.js";
 import { showPostCrawlMenu } from "./menus.js";
 import { displayCrawlSummary } from "./summary.js";
 import { showExtractedData } from "./viewer.js";
