+++ vite.config.mts
@@ -9,23 +9,21 @@ export default defineConfig({
 	},
 	build: {
 		target: "node18",
-		lib: {
-			entry: resolve(__dirname, "src/cli/index.ts"),
-			formats: ["es"],
-			fileName: "index",
-		},
 		minify: false,
 		rollupOptions: {
+			input: resolve(__dirname, "src/index.ts"),
 			external: [
 				"node:module",
 				"node:path",
 				"node:fs",
+				"node:fs/promises",
 				"node:os",
 				"node:process",
 				"node:stream",
 				"node:util",
 				"node:events",
 				"node:child_process",
+				"node:crypto",
 				"puppeteer",
 				"puppeteer-extra",
 				"puppeteer-extra-plugin-adblocker",
@@ -37,6 +35,7 @@ export default defineConfig({
 				entryFileNames: "[name].js",
 				chunkFileNames: "[name].js",
 				assetFileNames: "[name].[ext]",
+				banner: "#!/usr/bin/env node",
 			},
 		},
 	},
